Fix compilation error and test factories

+++ Game/core/src/GameObjects/Drawable.java
@@ -5,6 +5,6 @@ import com.badlogic.gdx.graphics.g2d.Sprite;
 public interface Drawable
 {
     public Sprite getDrawingSprite();
-    public double getX();
-    public double getY();
+    public float getX();
+    public float getY();
 }
